+++ game-engine/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./core/gameModule";
+export * from "./games/connect4/connect4Module";